requirements.txt
python-telegram-bot
google-generativeai
gspread
python-dotenv
oauth2client
flask